+++ SingularityS3Base/src/main/java/com/hubspot/singularity/s3/base/CacheCheck.java
@@ -1,0 +1,19 @@
+package com.hubspot.singularity.s3.base;
+
+public class CacheCheck {
+  private final CacheCheckResult cacheCheckResult;
+  private final String message;
+
+  public CacheCheck(CacheCheckResult cacheCheckResult, String message) {
+    this.cacheCheckResult = cacheCheckResult;
+    this.message = message;
+  }
+
+  public CacheCheckResult getCacheCheckResult() {
+    return cacheCheckResult;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+}
